LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/instantiation-error-5-expected.txt

CONSOLE MESSAGE: SyntaxError: Importing binding name 'something' is not found.
CONSOLE MESSAGE: SyntaxError: Importing binding name 'something' is not found.

FAIL Test that loading a graph in which a module is already errored results an error. assert_not_equals: Instantiation error objects for different root scripts got disallowed value "SyntaxError: Importing binding name 'something' is not found."
